reorganized app folder structure adding /user and /navigation folders along with index.js for export simplification on user.
