+++ src/app/navigation/AppRoutes.js
@@ -1,0 +1,95 @@
+import React, { Component } from 'react';
+import {
+  BrowserRouter as Router,
+  Route,
+  Link,
+  ReactDOM
+} from 'react-router-dom'
+import { connect } from "react-redux";
+import { bindActionCreators } from 'redux';
+// import user actions and destructure
+import { actions as userActions } from '../users';
+import NavMenu from './NavMenu'
+import RouteList from './RouteList'
+
+class AppRoutes extends Component {
+  constructor(props) {
+    super(props)
+
+    this.handleLogout = this.handleLogout.bind(this);
+  }
+
+  showSettings (event) {
+    event.preventDefault();
+  }
+
+  componentWillMount(){
+    // If user is not already authenticated
+    if (this.props.isAuth) return;
+    // Get token from local storage, quit if its null
+    let token = localStorage.getItem('jwtToken');
+  	if(!token || token === '') return;
+
+  	// fetch user from token (if server deems it's valid token)
+    this.props.meFromToken(token)
+    .then((response) => {
+      if (!response.error) {
+      	// reset token (possibly new token that was regenerated by the server)
+      	localStorage.setItem('jwtToken', response.payload.data.token);
+        this.props.meFromTokenSuccess(response.payload);
+      } else {
+      	localStorage.removeItem('jwtToken');//remove token from storage
+        this.props.meFromTokenFailure(response.payload);
+      }
+    });
+	 }
+
+   handleLogout() {
+     //Delete token from local storage
+     localStorage.removeItem('jwtToken');
+     // dispatch logout action
+     this.props.logoutUser();
+   }
+
+  render(){
+    return(
+      <Router>
+        <main id="outer-container">
+          <NavMenu handleLogout={this.handleLogout} />
+          <RouteList />
+        </main>
+      </Router>
+  )
+  }
+}
+
+function mapStateToProps(state) {
+  return {
+    	user: state.user.user,
+      isAuth: state.user.isAuth,
+      error: state.error,
+      loading: state.loading,
+    };
+}
+
+function mapDispatchToProps(dispatch) {
+  return bindActionCreators({ ...userActions }, dispatch);
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(AppRoutes);
+// export default AppRoutes;
+//REDUX MAGIC! This puts both of our functions into the Component's props and links them to dispatch
+// function mapDispatchToProps(dispatch){
+//   return bindActionCreators({ incrementCount, decrementCount }, dispatch);
+// }
+//
+// //MORE REDUX MAGIC! This function takes in all of our Application State and takes pieces of it and maps it
+// //to the Component's props.
+// function mapStateToProps(state) {
+//   return {
+//
+//   };
+// }
+//
+// export default connect(mapStateToProps, mapDispatchToProps)(AppRoutes)
+// // export default AppRoutes
