--- src/app/pages/LogInPage.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import LogInForm from '../ComponentsForm'
-
-class ContactPage extends React.Component {
-  submit = (values) => {
-    // print the form values to the console
-    console.log(values)
-  }
-  render() {
-    return (
-      <ContactForm onSubmit={this.submit} />
-    )
-  }
-}
